requirements.txt
streamlit==1.41.1
pinecone==5.0.1
python-dotenv==1.0.1
sentence-transformers==2.7.0
requests==2.32.3
